# feat: remove gunzip-maybe and use Node.js streams