--- a/package.json
+++ b/package.json
@@ -37,7 +37,6 @@
     "dockerfile-ast": "0.2.1",
     "elfy": "^1.0.0",
     "event-loop-spinner": "^2.0.0",
-    "gunzip-maybe": "^1.4.2",
     "mkdirp": "^1.0.4",
     "semver": "^7.3.4",
     "snyk-nodejs-lockfile-parser": "1.37.3",