- [x] Ready for review
- [x] Follows CONTRIBUTING rules
- [ ] Reviewed by Snyk internal team

#### What does this PR do?

While analyzing a performance issue, noticed that the CPU usage is pretty high. The flame graph / stack trace showed a subdependency of gunzip-maybe to be the main reason.
Switching to Node.js streams yields a slight performance improvement (tested with the latest Redis OCI archive):
- Before: 1.18s user 0.20s system 132% cpu 1.041 total
- After: 1.02s user 0.19s system 126% cpu 0.963 total